Migrated file.js to TS
